Creating a partial class for GameFuseUser Friends functions
