+++ com.gamefuse.sdk/Runtime/GameFuseUser.Friends.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: cc946179105389345a4df993e83279f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
